Move AudioPlayer to top-level
